requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
importlib-metadata==4.10.0
lxml==4.7.1
moviepy==1.0.3
numpy==1.22.0
paramiko==2.9.2
parse==1.19.0
Pillow==9.0.0
proglog==0.1.9
pycparser==2.21
pyee==8.2.2
PyNaCl==1.5.0
pyppeteer==1.0.2
pyquery==1.4.3
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
tornado==6.1
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
w3lib==1.22.0
websockets==10.1
zipp==3.7.0
